+++ package.js
@@ -1,11 +1,9 @@
-var both = ['client', 'server'];
 var client = 'client';
-var server = 'server';
 
 Package.describe({
     name: "wieldo:angular-formly-transformer",
     summary: "Automate configuration of fields in Angular-Formly",
-    version: "1.1.2",
+    version: "1.2.0",
 
     documentation: 'README.md',
     git: 'https://github.com/wieldo/angular-formly-transformer.git'
@@ -15,15 +13,15 @@ Package.onUse(function (api) {
 
     var packages = {
         use: [
+            'angular@1.0.0',
+            'angular:angular@1.4.7',
             'underscore@1.0.4',
             'pbastowski:angular-babel@1.0.2',
             'pbastowski:angular2-now@0.3.13',
             'wieldo:angular-formly@7.3.2'
         ],
         imply: [
-            'angular@1.0.0',
-            'angular:angular@1.4.7',
-            'wieldo:angular-formly@7.3.2'
+            'wieldo:angular-formly'
         ]
     };
 
@@ -34,20 +32,20 @@ Package.onUse(function (api) {
     api.imply(packages.imply);
 
     api.addFiles([
-        'lib/client/main.js',
-        'lib/client/formly-transformer.js'
+        'dist/formly-transformer.js'
     ], client);
 
 });
 
 Package.onTest(function(api) {
     api.use([
-        'underscore@1.0.4',
+        'pbastowski:angular-babel@1.0.2',
+        'pbastowski:angular2-now@0.3.13',
         'sanjo:jasmine@0.20.2',
         'velocity:helpers',
         'velocity:console-reporter',
         'angular:angular-mocks@1.4.7',
-        'wieldo:angular-formly-transformer@1.0.0'
+        'wieldo:angular-formly-transformer'
     ]);
     
     api.addFiles([