+++ tests/client/formly-transformer-spec.js
@@ -1,50 +1,70 @@
-describe('formlyTransformer', function () {
-    var formlyTransformer;
+describe('formlyTransformer', () => {
+    //
+    // vars
+    //
 
-    beforeEach(function () {
+    let formlyTransformer;
+
+    //
+    // helpers
+    //
+
+    function failRegister(values, errorMsg) {
+        values.forEach((value) => {
+            expect(() => {
+                formlyTransformer.register(value);
+            }).toThrowError(errorMsg);
+        });
+    }
+
+    function passRegister(values, errorMsg) {
+        values.forEach((value) => {
+            expect(() => {
+                formlyTransformer.register(value);
+            }).not.toThrowError(errorMsg);
+        });
+    }
+
+    //
+    // tests
+    //
+
+    beforeEach(() => {
         module('formlyTransformer');
         inject(function (_formlyTransformer_) {
             formlyTransformer = _formlyTransformer_;
         });
     });
 
-    it('should be empty on start', function () {
+    it('should be empty on start', () => {
         expect(formlyTransformer._transformers).toBeDefined();
         expect(formlyTransformer._transformers.length).toEqual(0);
         expect(formlyTransformer._transformers).toEqual([]);
     });
 
-    it('should have createError method which returns Error with prefixed message', function () {
-        var errorMsg = "[formlyTransformer] test";
+    it('should have createError method which returns Error with prefixed message', () => {
+        const errorMsg = "[formlyTransformer] test";
 
-        expect(function () {
+        expect(() => {
             throw formlyTransformer.createError('test');
         }).toThrowError(Error, errorMsg);
     });
 
-    it('should pass only functions', function () {
-        var errorMsg = "[formlyTransformer] Transformer is not a function";
-        var values = [undefined, false, true, 1, 0, -1, 's', '1', '0', '-1', 'true', 'false', {}, null, ['s']];
+    it('should pass only functions', () => {
+        const errorMsg = "[formlyTransformer] Transformer is not a function";
+        const values = [undefined, false, true, 1, 0, -1, 's', '1', '0', '-1', 'true', 'false', {}, null, ['s']];
 
-        expect(function () {
-            formlyTransformer.register(function () {
-            });
-        }).not.toThrowError(errorMsg);
+        passRegister([() => {
+        }], errorMsg);
 
-        _.each(values, function (value) {
-            expect(function () {
-                formlyTransformer.register(value);
-            }).toThrowError(errorMsg);
-        });
+        failRegister(values, errorMsg);
     });
 
-    it('should run registered transformers', function () {
-        var spy = jasmine.createSpy('spy');
-        var transformer = function (fields) {
-            spy(fields);
-        };
-        var transformed;
-        var expected = [
+    it('should run registered transformers', () => {
+        const spy = jasmine.createSpy('spy');
+        const transformer = (fields) => spy(fields);
+        let transformed;
+        const expected = [
             {
                 key: 'test-1'
             },
@@ -52,7 +72,7 @@ describe('formlyTransformer', function () {
                 key: 'test-2'
             }
         ];
-        var fields = [
+        const fields = [
             {
                 key: 'test-1',
                 transformers: {
@@ -75,12 +95,12 @@ describe('formlyTransformer', function () {
 
     });
 
-    it('should have createError method in transformer context', function () {
+    it('should have createError method in transformer context', () => {
         formlyTransformer.register(function () {
             throw this.createError('test');
         });
 
-        expect(function () {
+        expect(() => {
             formlyTransformer.run([{}, {}], {}, {}, {});
         }).toThrowError(Error, "[formlyTransformer] test");
     });