+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.2.0] - 2015-11-20
+### Added
+- Support for npm
+- Support for bower
+- Support for webpack
+- karma tests
+- coverage reports
+
 ## [1.1.2] - 2015-11-13
 ### Removed
 - es5-shim package
@@ -39,6 +47,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 0.0.1 - 2015-11-09
 
+[1.2.0]: https://github.com/wieldo/angular-formly-transformer/compare/v1.1.2...v1.2.0
 [1.1.2]: https://github.com/wieldo/angular-formly-transformer/compare/v1.1.1...v1.1.2
 [1.1.1]: https://github.com/wieldo/angular-formly-transformer/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/wieldo/angular-formly-transformer/compare/v1.0.0...v1.1.0